backend/requirements.txt
# ─── Web Framework ─────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ─── Socket.IO ─────────────────────────────────────────────────────────────────
python-socketio[asyncio]==5.11.3
websockets==12.0

# ─── Database ──────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
pgvector==0.3.2
alembic==1.13.1
psycopg2-binary==2.9.9

# ─── Redis ─────────────────────────────────────────────────────────────────────
redis[hiredis]==5.0.4

# ─── Auth ──────────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ─── AI / Computer Vision ──────────────────────────────────────────────────────
insightface==0.7.3
onnxruntime-gpu==1.18.0
# Fallback CPU runtime (uncomment if no GPU):
# onnxruntime==1.18.0
opencv-python-headless==4.9.0.80
mediapipe==0.10.14
Pillow==10.3.0
numpy==1.26.4
scikit-learn==1.4.2

# ─── Configuration ─────────────────────────────────────────────────────────────
pydantic-settings==2.2.1
python-dotenv==1.0.1

# ─── Utilities ─────────────────────────────────────────────────────────────────
structlog==24.1.0
prometheus-client==0.20.0
httpx==0.27.0
tenacity==8.2.3

# ─── OSINT Graph (optional) ───────────────────────────────────────────────
# datasets>=2.19.0         # HuggingFace datasets (uncomment for dataset ingestion)
# neo4j>=5.19.0            # Neo4j driver (uncomment for Neo4j graph backend)
